Add NMSUtils
