+++ scoreboard/spigot/src/main/java/net/apartium/cocoabeans/scoreboard/spigot/NMSUtils.java
@@ -17,7 +17,6 @@ import java.lang.invoke.MethodHandles;
 import java.lang.invoke.MethodType;
 import java.lang.reflect.Array;
 import java.lang.reflect.Field;
-import java.lang.reflect.Method;
 import java.lang.reflect.Modifier;
 import java.util.*;
 import java.util.function.Predicate;
@@ -614,37 +613,6 @@ import static net.apartium.cocoabeans.structs.MinecraftVersion.*;
     @FunctionalInterface
     public interface PacketConstructor {
         Object createInstance() throws Throwable;
-    }
-
-
-    private final static String
-            PACKAGE_NAME = Bukkit.getServer().getClass().getPackage().getName(),
-            SERVER_VERSION = PACKAGE_NAME.substring(PACKAGE_NAME.lastIndexOf(".") + 1);
-
-    private static final boolean USE_PACKAGE_WITH_VERSION = shouldUsePackageWithVersion();
-
-    private static boolean shouldUsePackageWithVersion() {
-        try {
-            Method getHandle = Bukkit.getServer().getClass().getMethod("getHandle");
-            return getHandle.getReturnType().getName().contains(SERVER_VERSION);
-        } catch (NoSuchMethodException e) {
-            String className = Bukkit.getServer().getClass().getName();
-            throw new RuntimeException("Could not find method getHandle of class %s".formatted(className), e);
-        }
-    }
-
-    public static String fixNMSFQDNForNonMappedFormat(String newFQDN) {
-        if (USE_PACKAGE_WITH_VERSION) {
-            String[] split = newFQDN.split("\\.");
-            String clazzName = split[split.length - 1];
-            return String.format("net.minecraft.server.%s.%s", SERVER_VERSION, clazzName);
-        }
-
-        return newFQDN;
-    }
-
-    public static String formatOBC(String path) {
-        return String.format("%s.%s", PACKAGE_NAME, path);
     }
